Use VERSION which requires an SES namespace
